More Modularization - Moved permission and moderator calculations to utils.js - Fixed retrieval of single guild configs
